Tweak docs, fix command and separate current tests
