+++ src/StackExchange.Redis/RedisLiterals.cs
@@ -53,6 +53,7 @@ namespace StackExchange.Redis
             CHANNELS = "CHANNELS",
             COPY = "COPY",
             COUNT = "COUNT",
+            DB = "DB",
             DESC = "DESC",
             DOCTOR = "DOCTOR",
             EX = "EX",
